Rename spring example
