+++ examples/spring-sleuth/README.md
@@ -27,7 +27,7 @@ You can run the example application via gradle. From the project root:
 ##### Build the executable JAR
 
 ```shell
-cd examples/spring/ && gradle bootJar
+cd examples/spring-sleuth/ && gradle bootJar
 ```
 
 ##### Run the executable JAR
@@ -35,7 +35,7 @@ cd examples/spring/ && gradle bootJar
 The JAR built from the previous command typically ends up in `build/libs` -
 
 ```shell
-java -jar build/libs/examples-spring-0.1.0-SNAPSHOT.jar
+java -jar build/libs/examples-spring-sleuth-0.1.0-SNAPSHOT.jar
 ```
 
 The application is now running. To generate traces, head to `http://localhost:8080` in your browser.
