+++ examples/spring-sleuth/src/main/java/com/google/cloud/opentelemetry/example/springsleuth/GreetingClient.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.opentelemetry.example.spring;
+package com.google.cloud.opentelemetry.example.springsleuth;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
