+++ examples/spring-sleuth/src/main/java/com/google/cloud/opentelemetry/example/springsleuth/Main.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.opentelemetry.example.spring;
+package com.google.cloud.opentelemetry.example.springsleuth;
 
 import com.google.cloud.opentelemetry.trace.TraceExporter;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
