+++ examples/spring-sleuth/src/main/resources/application.properties
